WIP refactor ArgoFileValidator
